+++ src/app/(admin)/_type/DetailTable/DetailTableItem.ts
