+++ src/app/(admin)/_type/DetailTable/DetailTableTitle.ts
